#1120 :truck: move `BaseOfflineLoading` to `components/common`-directory
