+++ wls-gui-wahllokalsystem/src/components/common/BaseOfflineLoading.vue
